abstract the MobileContentApiSessionInterceptor to utilize the GodToolsAccountManager
