+++ library/api/src/main/kotlin/org/cru/godtools/api/AuthApi.kt
@@ -2,7 +2,7 @@ package org.cru.godtools.api
 
 import org.ccci.gto.android.common.jsonapi.model.JsonApiObject
 import org.cru.godtools.api.model.AuthToken
-import retrofit2.Call
+import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
 
@@ -10,5 +10,5 @@ private const val PATH_AUTHENTICATE = "auth"
 
 interface AuthApi {
     @POST(PATH_AUTHENTICATE)
-    fun authenticate(@Body request: AuthToken.Request): Call<JsonApiObject<AuthToken>>
+    suspend fun authenticate(@Body request: AuthToken.Request): Response<JsonApiObject<AuthToken>>
 }
